chore(firstMile): fix typo - InitalizeClient => InitializeClient
