+++ src/homepageExperience/components/steps/python/InitializeClient.tsx
@@ -4,47 +4,51 @@ import {useSelector} from 'react-redux'
 import CodeSnippet from 'src/shared/components/CodeSnippet'
 import {event} from 'src/cloud/utils/reporting'
 
+import {getOrg} from 'src/organizations/selectors'
 import {getMe} from 'src/me/selectors'
 
 const logCopyCodeSnippet = () => {
-  event('firstMile.nodejsWizard.initializeClient.code.copied')
+  event('firstMile.pythonWizard.initializeClient.code.copied')
 }
 
-export const InitalizeClient = () => {
+export const InitializeClient = () => {
+  const org = useSelector(getOrg)
   const me = useSelector(getMe)
 
   const url =
     me.quartzMe?.clusterHost || 'https://us-west-2-1.aws.cloud2.influxdata.com/'
 
-  const codeSnippet = `repl.repl.ignoreUndefined=true
+  const pythonCode = `import influxdb_client, os, time
+from influxdb_client import InfluxDBClient, Point, WritePrecision
+from influxdb_client.client.write_api import SYNCHRONOUS
 
-const {InfluxDB, Point} = require('@influxdata/influxdb-client')
+token = os.environ.get("INFLUXDB_TOKEN")
+org = "${org.name}"
+url = "${url}"
 
-const token = process.env.INFLUXDB_TOKEN
-const url = '${url}'
-
-const client = new InfluxDB({url, token})`
+client = influxdb_client.InfluxDBClient(url=url, token=token, org=org)
+`
 
   return (
     <>
       <h1>Initialize Client</h1>
       <p>
-        Run this command in your terminal to open the interactive Nodejs shell:
+        Run this command in your terminal to open the interactive Python shell:
       </p>
-      <CodeSnippet text="node" language="properties" />
+      <CodeSnippet text="python3" language="properties" />
       <p style={{marginTop: '40px'}}>
-        Paste the following code after the prompt (>) and press Enter.
+        Paste the following code after the prompt (>>>) and press Enter.
       </p>
       <CodeSnippet
-        text={codeSnippet}
+        text={pythonCode}
         onCopy={logCopyCodeSnippet}
-        language="javascript"
+        language="python"
       />
       <p style={{marginTop: '42px'}}>
         Here, we initialize the token, organization info, and server url that
         are needed to set up the initial connection to InfluxDB. The client
         connection is then established with the{' '}
-        <code className="homepage-wizard--code-highlight">InfluxDB</code>{' '}
+        <code className="homepage-wizard--code-highlight">InfluxDBClient</code>{' '}
         initialization.
       </p>
     </>
